refactor: extract redirect middleware to core/redirects, business logic to services/redirect_service
Address Copilot PR review comments:
- Move middleware out of main.py into core/redirects.py (pure ASGI, no service imports)
- Extract URL resolution logic to services/redirect_service.py
- Inject resolver via constructor to respect core -> services dependency direction
- Replace raw query string logging with has_query/query_length (security)
- Add graceful exception handling in middleware (no 500 on resolver failure)
- Mock alias data in tests for deterministic, YAML-independent test runs
